feat: add response error codes
