+++ src/Triggers/TriggerStore.js
@@ -1,0 +1,208 @@
+import logger from '../logger';
+
+export const Types = {
+  beforeLogin: 'beforeLogin',
+  afterLogin: 'afterLogin',
+  afterLogout: 'afterLogout',
+  beforeSave: 'beforeSave',
+  afterSave: 'afterSave',
+  beforeDelete: 'beforeDelete',
+  afterDelete: 'afterDelete',
+  beforeFind: 'beforeFind',
+  afterFind: 'afterFind',
+  beforeConnect: 'beforeConnect',
+  beforeSubscribe: 'beforeSubscribe',
+  afterEvent: 'afterEvent',
+};
+
+const baseStore = function () {
+  const Validators = Object.keys(Types).reduce(function (base, key) {
+    base[key] = {};
+    return base;
+  }, {});
+  const Functions = {};
+  const Jobs = {};
+  const LiveQuery = [];
+  const Triggers = Object.keys(Types).reduce(function (base, key) {
+    base[key] = {};
+    return base;
+  }, {});
+
+  return Object.freeze({
+    Functions,
+    Jobs,
+    Validators,
+    Triggers,
+    LiveQuery,
+  });
+};
+
+function validateClassNameForTriggers(className, type) {
+  if (type == Types.beforeSave && className === '_PushStatus') {
+    // _PushStatus uses undocumented nested key increment ops
+    // allowing beforeSave would mess up the objects big time
+    // TODO: Allow proper documented way of using nested increment ops
+    throw 'Only afterSave is allowed on _PushStatus';
+  }
+  if ((type === Types.beforeLogin || type === Types.afterLogin) && className !== '_User') {
+    // TODO: check if upstream code will handle `Error` instance rather
+    // than this anti-pattern of throwing strings
+    throw 'Only the _User class is allowed for the beforeLogin and afterLogin triggers';
+  }
+  if (type === Types.afterLogout && className !== '_Session') {
+    // TODO: check if upstream code will handle `Error` instance rather
+    // than this anti-pattern of throwing strings
+    throw 'Only the _Session class is allowed for the afterLogout trigger.';
+  }
+  if (className === '_Session' && type !== Types.afterLogout) {
+    // TODO: check if upstream code will handle `Error` instance rather
+    // than this anti-pattern of throwing strings
+    throw 'Only the afterLogout trigger is allowed for the _Session class.';
+  }
+  return className;
+}
+
+const _triggerStore = {};
+
+const Category = {
+  Functions: 'Functions',
+  Validators: 'Validators',
+  Jobs: 'Jobs',
+  Triggers: 'Triggers',
+};
+
+function getStore(category, name, applicationId) {
+  const invalidNameRegex = /['"`]/;
+  if (invalidNameRegex.test(name)) {
+    // Prevent a malicious user from injecting properties into the store
+    return {};
+  }
+
+  const path = name.split('.');
+  path.splice(-1); // remove last component
+  applicationId = applicationId || Parse.applicationId;
+  _triggerStore[applicationId] = _triggerStore[applicationId] || baseStore();
+  let store = _triggerStore[applicationId][category];
+  for (const component of path) {
+    store = store[component];
+    if (!store) {
+      return {};
+    }
+  }
+  return store;
+}
+
+function add(category, name, handler, applicationId) {
+  const lastComponent = name.split('.').splice(-1);
+  const store = getStore(category, name, applicationId);
+  if (store[lastComponent]) {
+    logger.warn(
+      `Warning: Duplicate cloud functions exist for ${lastComponent}. Only the last one will be used and the others will be ignored.`
+    );
+  }
+  store[lastComponent] = handler;
+}
+
+function remove(category, name, applicationId) {
+  const lastComponent = name.split('.').splice(-1);
+  const store = getStore(category, name, applicationId);
+  delete store[lastComponent];
+}
+
+function get(category, name, applicationId) {
+  const lastComponent = name.split('.').splice(-1);
+  const store = getStore(category, name, applicationId);
+  return store[lastComponent];
+}
+
+export function addFunction(functionName, handler, validationHandler, applicationId) {
+  add(Category.Functions, functionName, handler, applicationId);
+  add(Category.Validators, functionName, validationHandler, applicationId);
+}
+
+export function addJob(jobName, handler, applicationId) {
+  add(Category.Jobs, jobName, handler, applicationId);
+}
+
+export function addTrigger(type, className, handler, applicationId, validationHandler) {
+  validateClassNameForTriggers(className, type);
+  add(Category.Triggers, `${type}.${className}`, handler, applicationId);
+  add(Category.Validators, `${type}.${className}`, validationHandler, applicationId);
+}
+
+export function addLiveQueryEventHandler(handler, applicationId) {
+  applicationId = applicationId || Parse.applicationId;
+  _triggerStore[applicationId] = _triggerStore[applicationId] || baseStore();
+  _triggerStore[applicationId].LiveQuery.push(handler);
+}
+
+export function runLiveQueryEventHandlers(data, applicationId = Parse.applicationId) {
+  if (!_triggerStore || !_triggerStore[applicationId] || !_triggerStore[applicationId].LiveQuery) {
+    return;
+  }
+  _triggerStore[applicationId].LiveQuery.forEach(handler => handler(data));
+}
+
+export function removeFunction(functionName, applicationId) {
+  remove(Category.Functions, functionName, applicationId);
+}
+
+export function removeTrigger(type, className, applicationId) {
+  remove(Category.Triggers, `${type}.${className}`, applicationId);
+}
+
+export function _unregisterAll() {
+  Object.keys(_triggerStore).forEach(appId => delete _triggerStore[appId]);
+}
+
+export function triggerExists(className: string, type: string, applicationId: string): boolean {
+  return getTrigger(className, type, applicationId) != undefined;
+}
+
+export function getTrigger(className, triggerType, applicationId) {
+  if (!applicationId) {
+    throw 'Missing ApplicationID';
+  }
+  return get(Category.Triggers, `${triggerType}.${className}`, applicationId);
+}
+
+export function getFunction(functionName, applicationId) {
+  return get(Category.Functions, functionName, applicationId);
+}
+
+export function getValidator(functionName, applicationId) {
+  return get(Category.Validators, functionName, applicationId);
+}
+
+export function getJob(jobName, applicationId) {
+  return get(Category.Jobs, jobName, applicationId);
+}
+
+export function getJobs(applicationId) {
+  var manager = _triggerStore[applicationId];
+  if (manager && manager.Jobs) {
+    return manager.Jobs;
+  }
+  return undefined;
+}
+
+export function getFunctionNames(applicationId) {
+  const store =
+    (_triggerStore[applicationId] && _triggerStore[applicationId][Category.Functions]) || {};
+  const functionNames = [];
+  const extractFunctionNames = (namespace, store) => {
+    Object.keys(store).forEach(name => {
+      const value = store[name];
+      if (namespace) {
+        name = `${namespace}.${name}`;
+      }
+      if (typeof value === 'function') {
+        functionNames.push(name);
+      } else {
+        extractFunctionNames(name, value);
+      }
+    });
+  };
+  extractFunctionNames(null, store);
+  return functionNames;
+}
